2017053_1411

+++ GUI_Layout_1/GUI_Layout_1/frame_daq_measure.py
@@ -77,10 +77,10 @@ class daq_measure_frame(tk.Frame):
         self.running = not self.running
 
     def start(self):
-        self.daq9211.measureAll(int(self.interval.get()))
+        self.daq9211.measureAll(float(self.interval.get()))
         t = threading.Thread(target = self.animation_target)
         t.start()
-        #self.ani = animation.FuncAnimation(self.fig, self.update_graph, interval = int(self.interval.get())*1000 + 1, repeat=False)
+        #self.ani = animation.FuncAnimation(self.fig, self.update_graph, interval = float(self.interval.get())*1000 + 1, repeat=False)
         self.ani = True
         self.running = True
 
@@ -107,7 +107,7 @@ class daq_measure_frame(tk.Frame):
     def animation_target(self):
         self.stopgraph_event.clear()
         while(not self.stopgraph_event.is_set()):
-            time.sleep(int(self.interval.get()))
+            time.sleep(float(self.interval.get()))
             self.update_graph()
         self.stopgraph_event.set() #once animation stops, reset the stop event to trigger again
